Move TensorflowTileDBSparseDataset to readers.tensorflow, change generators from classmethods to plain functions and move to separate module
